model/requirements.txt
firebase_admin==6.5.0
PyQt5==5.15.11
PyQt5_sip==12.13.0
python_bcrypt==0.3.2
